fix: remediate pkg/service file count for wire composition PR
Consolidate composition seams into factory_editable_definition.go and move
compose equivalence coverage to factoryvalidationtests per pkg-file-count gate.
